etl_pipeline_b2_and_b3/requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-celery

snowflake-connector-python

langchain
langchain-core
langchain-xai
langgraph

sqlglot

